add continue learning setting

+++ learn_log.py
@@ -5,12 +5,13 @@ import matplotlib.pyplot as plt
 class MetricLogger():
     def __init__(self, save_dir):
         self.save_log = save_dir / "log"
-        with open(self.save_log, "w") as f:
-            f.write(
-                f"{'Episode':>8}{'Step':>8}{'Epsilon':>10}{'MeanReward':>15}"
-                f"{'MeanLength':>15}{'MeanLoss':>15}{'MeanQValue':>15}"
-                f"{'TimeDelta':>15}{'Time':>20}\n"
-            )
+        if not self.save_log.is_file():
+            with open(self.save_log, "w") as f:
+                f.write(
+                    f"{'Episode':>8}{'Step':>8}{'Epsilon':>10}{'MeanReward':>15}"
+                    f"{'MeanLength':>15}{'MeanLoss':>15}{'MeanQValue':>15}"
+                    f"{'TimeDelta':>15}{'Time':>20}\n"
+                )
         self.ep_rewards_plot = save_dir / "reward_plot.jpg"
         self.ep_lengths_plot = save_dir / "length_plot.jpg"
         self.ep_avg_losses_plot = save_dir / "loss_plot.jpg"
